refactor: move browser to dedicated directory
